+++ Changelog.rst
@@ -1,10 +1,8 @@
 Changelog
 =========
 
-0.4.0:
- - change ``Client.check_module_release`` to ``Client.check_package_release``
-
 0.3.3:
+ - change ``Client.check_module_release`` to ``Client.check_package_release``
  - support output filename with deep directory
  - Always sort cnt files to avoid merge error (7979473)
  - Fix typos