+++ Changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.3.3:
+0.3.3 (2017-03-17):
  - change ``network`` to ``NETWORK`` in ``header.py``
  - add wildcard support to ``win32.merge``
  - change ``Client.check_module_release`` to ``Client.check_package_release``